tools/requirements.txt
PyYAML==6.0.3
minio==7.2.20
protobuf==7.34.1
pulsar-client==3.11.0
websocket-client==1.9.0
